Task 4. Refactoring, moved not kafka related logic from consumers to dedicated services
