+++ t1-java/src/main/java/ru/t1/java/demo/kafka/consumer/TransactionResultConsumer.java
@@ -10,21 +10,14 @@ import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.stereotype.Component;
 import ru.t1.java.demo.exception.IllegalOperationTypeException;
 import ru.t1.java.demo.kafka.message.TransactionResultMessage;
-import ru.t1.java.demo.model.Account;
-import ru.t1.java.demo.model.Transaction;
-import ru.t1.java.demo.model.enums.AccountStatus;
-import ru.t1.java.demo.model.enums.OperationType;
-import ru.t1.java.demo.model.enums.TransactionStatus;
-import ru.t1.java.demo.service.AccountService;
-import ru.t1.java.demo.service.TransactionService;
+import ru.t1.java.demo.service.TransactionResultService;
 
 @Slf4j
 @RequiredArgsConstructor
 @Component
 public class TransactionResultConsumer {
 
-    private final TransactionService transactionService;
-    private final AccountService accountService;
+    private final TransactionResultService transactionResultService;
 
     @KafkaListener(groupId = "${t1.kafka.consumer.transaction-result-id}",
             topics = "${t1.kafka.topic.transaction_result}",
@@ -34,36 +27,7 @@ public class TransactionResultConsumer {
         log.info("TransactionResult consumer: new message from {} topic", topic);
 
         try {
-            TransactionStatus status = transactionResultMessage.getStatus();
-            Transaction transaction = new Transaction();
-
-            switch (status) {
-                case ACCEPTED -> {
-                    transaction.setStatus(TransactionStatus.ACCEPTED);
-                    transactionService.updateTransaction(transactionResultMessage.getTransactionId(), transaction);
-                }
-                case REJECTED -> {
-                    if (transactionResultMessage.getOperationType() == OperationType.INCOMING)
-                        accountService.decreaseBalance(transactionResultMessage.getAccountId(), transactionResultMessage.getTransactionAmount());
-                    else if (transactionResultMessage.getOperationType() == OperationType.OUTGOING)
-                        accountService.increaseBalance(transactionResultMessage.getAccountId(), transactionResultMessage.getTransactionAmount());
-                    else
-                        throw new IllegalOperationTypeException("Unknown operation type: " + transactionResultMessage.getOperationType());
-
-                    transaction.setStatus(TransactionStatus.REJECTED);
-                    transactionService.updateTransaction(transactionResultMessage.getTransactionId(), transaction);
-                }
-                case BLOCKED -> {
-                    Account account = Account.builder().status(AccountStatus.BLOCKED).build();
-                    accountService.updateAccount(transactionResultMessage.getAccountId(), account);
-
-                    transaction.setStatus(TransactionStatus.BLOCKED);
-                    transactionService.updateTransaction(transactionResultMessage.getTransactionId(), transaction);
-                }
-                default ->
-                    throw new IllegalArgumentException("Unknown transaction status: " + status);
-            }
-
+            transactionResultService.processTransactionResult(transactionResultMessage);
         } catch (IllegalOperationTypeException illegalOperationTypeException) {
             log.error("TransactionResult consumer: unknown operation type: {}", transactionResultMessage.getOperationType(), illegalOperationTypeException);
         } catch (IllegalArgumentException illegalArgumentException) {
